+++ forge/src/main/kotlin/com/github/kittinunf/forge/core/JSON.kt
@@ -1,6 +1,6 @@
-package com.github.kttinunf.forge.core
+package com.github.kittinunf.forge.core
 
-import com.github.kttinunf.forge.extension.asSequence
+import com.github.kittinunf.forge.extension.asSequence
 import org.json.JSONArray
 import org.json.JSONObject
