+++ forge/src/test/kotlin/JSONMappingObjectTest.kt
@@ -1,15 +1,15 @@
 import com.github.kttinunf.forge.Forge
-import com.github.kttinunf.forge.core.Deserializable
-import com.github.kttinunf.forge.core.EncodedResult
-import com.github.kttinunf.forge.core.JSON
-import com.github.kttinunf.forge.core.apply
-import com.github.kttinunf.forge.core.at
-import com.github.kttinunf.forge.core.list
-import com.github.kttinunf.forge.core.map
-import com.github.kttinunf.forge.core.maybeAt
-import com.github.kttinunf.forge.function.toDate
-import com.github.kttinunf.forge.util.create
-import com.github.kttinunf.forge.util.curry
+import com.github.kittinunf.forge.core.Deserializable
+import com.github.kittinunf.forge.core.EncodedResult
+import com.github.kittinunf.forge.core.JSON
+import com.github.kittinunf.forge.core.apply
+import com.github.kittinunf.forge.core.at
+import com.github.kittinunf.forge.core.list
+import com.github.kittinunf.forge.core.map
+import com.github.kittinunf.forge.core.maybeAt
+import com.github.kittinunf.forge.function.toDate
+import com.github.kittinunf.forge.util.create
+import com.github.kittinunf.forge.util.curry
 import org.hamcrest.CoreMatchers.equalTo
 import org.hamcrest.CoreMatchers.notNullValue
 import org.hamcrest.CoreMatchers.nullValue
